+++ src/main/java/com/yuuko/core/Configuration.java
@@ -25,7 +25,7 @@ import java.util.Set;
 public class Configuration {
     private static final Logger log = LoggerFactory.getLogger(Configuration.class);
 
-    public static final String VERSION = "14-03-2019_1";
+    public static final String VERSION = "17-03-2019_1";
     public static String AUTHOR;
     public static String AUTHOR_WEBSITE;
     public static String SUPPORT_GUILD;