+++ src/main/java/com/yuuko/core/commands/audio/commands/PlayCommand.java
@@ -15,7 +15,6 @@ import com.yuuko.core.utilities.MessageHandler;
 import com.yuuko.core.utilities.TextUtility;
 import lavalink.client.io.Link;
 import net.dv8tion.jda.core.EmbedBuilder;
-import net.dv8tion.jda.core.entities.MessageChannel;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 
 import java.util.List;
@@ -46,7 +45,7 @@ public class PlayCommand extends Command {
             manager.player.setPaused(false);
 
             if(command[1].startsWith("https://") || command[1].startsWith("http://")) {
-                loadAndPlay(manager, e.getChannel(), command[1], e);
+                loadAndPlay(manager, e, command[1]);
 
             } else {
                 String trackId = YouTubeSearchHandler.search(command[1]);
@@ -55,7 +54,7 @@ public class PlayCommand extends Command {
                     EmbedBuilder embed = new EmbedBuilder().setTitle("Those search parameters failed to return a result.");
                     MessageHandler.sendMessage(e, embed.build());
                 } else {
-                    loadAndPlay(manager, e.getChannel(), trackId, e);
+                    loadAndPlay(manager, e, trackId);
                 }
             }
         }
@@ -64,11 +63,12 @@ public class PlayCommand extends Command {
 
     /**
      * Loads a track from a given url and plays it if possible, else adds it to the queue.
-     * @param manager ; GuildAudioManager.
-     * @param channel ; MessageChannel.
-     * @param url ; TrackUrl.
+     *
+     * @param manager; GuildAudioManager.
+     * @param e; MessageReceivedEvent.
+     * @param url; TrackUrl.
      */
-    private void loadAndPlay(GuildAudioManager manager, final MessageChannel channel, String url, MessageReceivedEvent e) {
+    private void loadAndPlay(GuildAudioManager manager, MessageReceivedEvent e, String url) {
         final String trackUrl;
 
         if(url.startsWith("<") && url.endsWith(">")) {
@@ -96,7 +96,7 @@ public class PlayCommand extends Command {
                             .addField("Channel", track.getInfo().author, true)
                             .addField("Position in queue", manager.scheduler.queue.size() + "", false)
                             .setFooter(Configuration.STANDARD_STRINGS[1] + e.getMember().getEffectiveName(), e.getAuthor().getEffectiveAvatarUrl());
-                    MessageHandler.sendMessage(channel, embed.build());
+                    MessageHandler.sendMessage(e, embed.build());
                 } catch(Exception ex) {
                     log.error("An error occurred while running the {} class, message: {}", this, ex.getMessage(), ex);
                 }
@@ -106,7 +106,7 @@ public class PlayCommand extends Command {
             public void playlistLoaded(AudioPlaylist playlist) {
                 try {
                     EmbedBuilder embed = new EmbedBuilder().setTitle("Adding **" + playlist.getTracks().size() + "** tracks to queue from playlist: **" + playlist.getName() + "**");
-                    MessageHandler.sendMessage(channel, embed.build());
+                    MessageHandler.sendMessage(e, embed.build());
 
                     List<AudioTrack> tracks = playlist.getTracks();
                     for(AudioTrack track: tracks) {
@@ -124,13 +124,13 @@ public class PlayCommand extends Command {
             @Override
             public void noMatches() {
                 EmbedBuilder embed = new EmbedBuilder().setTitle("No matches found using that parameter.");
-                MessageHandler.sendMessage(channel, embed.build());
+                MessageHandler.sendMessage(e, embed.build());
             }
 
             @Override
             public void loadFailed(FriendlyException exception) {
                 EmbedBuilder embed = new EmbedBuilder().setTitle("Loading failed: " + exception.getMessage());
-                MessageHandler.sendMessage(channel, embed.build());
+                MessageHandler.sendMessage(e, embed.build());
             }
         });
     }